add user controller
