+++ back-end/bloco-29-arquitetura-solid-e-orm/dia-2-orm-interface-da-aplicacao-com-o-banco-de-daddos/exercicios-fixacao/app-with-sequelize/app.js
@@ -1,0 +1,9 @@
+const express = require('express');
+
+const app = express();
+
+const userRouter = require('./routers/userRouter');
+
+app.use('/user', userRouter);
+
+module.exports = app;
