+++ back-end/bloco-29-arquitetura-solid-e-orm/dia-2-orm-interface-da-aplicacao-com-o-banco-de-daddos/exercicios-fixacao/app-with-sequelize/controllers/userController.js
@@ -1,0 +1,16 @@
+const { User } = require('../models');
+
+const findAll = async (_req, res) => {
+    try{
+        const users = await User.findAll();
+
+        return res.status(200).json(users);
+    }catch(e) {
+        console.log(e.message)
+        res.status(500).json({message: 'algo deu errado'});
+    }
+};
+
+module.exports = {
+    findAll,
+};
