+++ back-end/bloco-29-arquitetura-solid-e-orm/dia-2-orm-interface-da-aplicacao-com-o-banco-de-daddos/exercicios-fixacao/app-with-sequelize/index.js
@@ -1,20 +1,3 @@
-const express = require('express');
-
-const app = express();
-
-const router = express.Router();
-
-const { User } = require('./models');
-
-app.get('/', async (_req, res) => {
-    try{
-        const users = await User.findAll();
-
-        return res.status(200).json(users);
-    }catch(e) {
-        console.log(e.message)
-        res.status(500).json({message: 'algo deu errado'});
-    }
-});
+const app = require('./app');
 
 app.listen(3000, () => console.log('Ouvindo na porta 3000'));
