+++ back-end/bloco-29-arquitetura-solid-e-orm/dia-2-orm-interface-da-aplicacao-com-o-banco-de-daddos/exercicios-fixacao/app-with-sequelize/package.json
@@ -4,6 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
+    "dev": "nodemon index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
@@ -13,6 +14,7 @@
     "body-parser": "^1.19.1",
     "express": "^4.17.2",
     "mysql2": "^2.3.3",
+    "nodemon": "^2.0.15",
     "sequelize": "^6.12.5",
     "sequelize-cli": "^6.3.0"
   }
