+++ back-end/bloco-29-arquitetura-solid-e-orm/dia-2-orm-interface-da-aplicacao-com-o-banco-de-daddos/exercicios-fixacao/app-with-sequelize/routers/userRouter.js
@@ -1,0 +1,9 @@
+const express = require('express');
+
+const userController = require('../controllers/userController');
+
+const route = express.Router();
+
+route.get('/', userController.findAll)
+
+module.exports = route;
